Use config.local.yaml
Before
=====

There is a default yaml file for configuration, but it is
checked into version control.

To change the default, you have to either pass the configuration every
time or have a modified local file.

After
=====

If you have a config.local.yaml (not in version control), that will
be the default configuration. Otherwise, it will default to
config.yaml